requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Mako==1.3.7
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.8
redis==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
pytz==2024.2
